requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
playwright==1.32.1
beautifulsoup4==4.10.0
geopy==2.4.1
newsapi-python==0.2.7